refactor : List<Integer> numbers => LottoNumbers 일급 컬렉션으로 변경
