+++ src/main/java/lotto/ui/Printer.java
@@ -16,7 +16,7 @@ public class Printer {
     public static void printLottoNumbers(List<Lotto> lottos) {
         System.out.printf("%d개를 구매했습니다.%n", lottos.size());
         lottos.forEach(lotto -> {
-            System.out.println(lotto.getNumbers().toString());
+            System.out.println(lotto.getLottoNumbers().toString());
         });
     }
